+++ Source/Scene/Primitive.js
@@ -204,7 +204,7 @@ define([
          * by {@link Transforms.eastNorthUpToFixedFrame}.
          *
          * <p>
-         * If the model matrix is changed after creation, it only affects primitives with one instance and only in 3D mode.
+         * This property is only supported in 3D mode.
          * </p>
          *
          * @type Matrix4
@@ -706,7 +706,7 @@ define([
      *
      * @exception {DeveloperError} All instance geometries must have the same primitiveType.
      * @exception {DeveloperError} Appearance and material have a uniform with the same name.
-     * @exception {DeveloperError} Primitive.modelMatrix is only supported in 3D mode and only for single geometry instances.
+     * @exception {DeveloperError} Primitive.modelMatrix is only supported in 3D mode.
      */
     Primitive.prototype.update = function(context, frameState, commandList) {
         if (((!defined(this.geometryInstances)) && (this._va.length === 0)) ||
@@ -1215,19 +1215,12 @@ define([
             attributes.length = 0;
         }
 
-        var modelMatrix;
-        if ((this._numberOfInstances > 1 && !this._validModelMatrix) || frameState.mode !== SceneMode.SCENE3D) {
-            modelMatrix = Matrix4.IDENTITY;
-
-            //>>includeStart('debug', pragmas.debug);
-            if (!Matrix4.equals(this.modelMatrix, Matrix4.IDENTITY)) {
-                throw new DeveloperError('Primitive.modelMatrix is only supported in 3D mode and only for single geometry instances.');
-            }
-            //>>includeEnd('debug');
-            
-        } else {
-            modelMatrix = this.modelMatrix;
+        var modelMatrix = this.modelMatrix;
+        //>>includeStart('debug', pragmas.debug);
+        if (frameState.mode !== SceneMode.SCENE3D && !Matrix4.equals(modelMatrix, Matrix4.IDENTITY)) {
+            throw new DeveloperError('Primitive.modelMatrix is only supported in 3D mode.');
         }
+        //>>includeEnd('debug');
 
         if (!Matrix4.equals(modelMatrix, this._modelMatrix)) {
             Matrix4.clone(modelMatrix, this._modelMatrix);