+++ CHANGES.md
@@ -2,6 +2,13 @@ Change Log
 ==========
 
 ### 1.8 -2015-04-01
+* Breaking changes
+  *
+* Deprecated
+  *
+* An exception is now thrown if `Primitive.modelMatrix` is not the identity matrix when in in 2D or Columbus View.
+
+### 1.8 -2015-04-01
 
 * Breaking changes
   * Removed the `eye`, `target`, and `up` parameters to `Camera.lookAt` which were deprecated in Cesium 1.6. Use the `target` and `offset`.
@@ -28,7 +35,6 @@ Change Log
 * Added new construction options to `CesiumWidget` and `Viewer`, for `skyBox`, `skyAtmosphere`, and `globe`.
 * Fixed a bug that prevented Cesium from working in browser configurations that explicitly disabled localStorage, such as Safari's private browsing mode.
 * Cesium is now tested using Jasmine 2.2.0.
-* Exception will be thrown on updating Primitive.modelMatrix in 2D or Columbus View or if it has more than one instance.
 
 ### 1.7.1 - 2015-03-06